refactor(Message): ErrorMessage SuccessMessage 구분
