+++ src/main/java/gift/constant/SuccessMessage.java
@@ -1,0 +1,8 @@
+package gift.constant;
+
+public class SuccessMessage {
+    public static final String ADD_SUCCESS_MSG = "상품 추가 성공";
+    public static final String UPDATE_SUCCESS_MSG = "상품 수정 성공";
+    public static final String DELETE_SUCCESS_MSG = "상품 삭제 성공";
+    public static final String ADD_OPTION_SUCCESS_MSG = "상품 옵션 추가 성공";
+}
