+++ src/main/java/gift/dto/request/AddProductRequest.java
@@ -2,7 +2,7 @@ package gift.dto.request;
 
 import jakarta.validation.constraints.*;
 
-import static gift.constant.Message.*;
+import static gift.constant.ErrorMessage.*;
 
 public record AddProductRequest(
