+++ src/main/java/gift/dto/request/OrderRequest.java
@@ -4,8 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-import static gift.constant.Message.POSITIVE_NUMBER_REQUIRED_MSG;
-import static gift.constant.Message.REQUIRED_FIELD_MSG;
+import static gift.constant.ErrorMessage.POSITIVE_NUMBER_REQUIRED_MSG;
+import static gift.constant.ErrorMessage.REQUIRED_FIELD_MSG;
 
 public record OrderRequest(
         @NotNull(message = REQUIRED_FIELD_MSG)
