+++ src/main/java/gift/service/ProductService.java
@@ -18,7 +18,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-import static gift.constant.Message.*;
+import static gift.constant.SuccessMessage.*;
 import static gift.exception.ErrorCode.DATA_NOT_FOUND;
 
 @Service
