+++ src/main/java/gift/service/WishListService.java
@@ -17,8 +17,8 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-import static gift.constant.Message.ADD_SUCCESS_MSG;
-import static gift.constant.Message.DELETE_SUCCESS_MSG;
+import static gift.constant.SuccessMessage.ADD_SUCCESS_MSG;
+import static gift.constant.SuccessMessage.DELETE_SUCCESS_MSG;
 import static gift.exception.ErrorCode.DATA_NOT_FOUND;
 
 @Service
